+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to Divisor will be documented in this file.
 
 > **NOTE**: The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 > and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+> Major and Minor updates will get entries, patches may not get mentions.
 
 ## [0.1.0] - 2023/07/20